[Update] Add DebugDust class for generating log file
This commit adds a new file "DebugDust.hpp" that contains the implementation of a DebugDust class. The class provides a static function 'GenerateLogFile' that takes a path as input and generates a file containing the logs. The logs are generated by concatenating the type and log message from each log entry.

Additionally, a new function 'PrintLog' has been added to the debug.hpp file. This function prints logs based on the type, time, fileName, function, message, and line provided. The logs are only printed if the type is not LOG_DEBUG and if the code is not running in the ENGINE_EDITOR environment.

In the engine.hpp file, the 'lastRenderClearTime' variable has been renamed to 'lastEngineCleanUp', and 'renderClearCoolDown' has been renamed to 'engineCleanup'. The function 'GenerateLogFile' is called after the engine cleanup process, and the file is named "logs.log".

In the render.hpp file, the 'generateGLTexture' and 'generateGLMesh' functions now also generate warning logs for unused textures and meshes, respectively.

These modifications aim to improve debugging capabilities and provide better management of logs and resources in the engine.