Refactor external.cmake and include PrettyEngine/editor.hpp
This commit refactors the external.cmake file to remove unnecessary options for building dependencies. It also includes the necessary directories for glad, glm, and stb. Additionally, it removes unused options for building OpenAL Soft.

In the PrettyEngine/editor.hpp file, a button is added to remove entities from the editor. When pressed, the corresponding entity is removed from the world and the selected entities list is cleared.