Refactor custom and components libraries
The modifications in this commit refactor the custom and components libraries. In the custom.cmake file, the custom library is changed from being built as a shared library to a regular library. Similarly, in the external.cmake file, the external library is changed from being built as a shared library to a regular library as well. This change likely impacts the dependencies and linking of these libraries.