Update build configurations and remove unnecessary package installations
The commit modifies the build configurations of the project by adjusting the matrix for `os` and `c_compiler`. The `c_compiler` for the `ubuntu-latest` environment is removed. The commit also removes the installation of the following packages on `ubuntu-latest`: `libxrandr-dev`, `libxinerama-dev`, and `libxcursor-dev`. These package installations are no longer necessary.